Move ssh constants to correct file
