+++ packages/api/ssh/models.go
@@ -1,15 +1,15 @@
 package api
 
 import (
-	"github.com/infisical/go-sdk/packages/models"
+	"github.com/infisical/go-sdk/packages/util"
 )
 
 type SignSshPublicKeyV1Request struct {
 	ProjectID    string                  `json:"projectId"`
 	TemplateName string                  `json:"templateName"`
 	PublicKey    string                  `json:"publicKey"`
-	KeyAlgorithm models.CertKeyAlgorithm `json:"keyAlgorithm,omitempty"`
-	CertType     models.SshCertType      `json:"certType,omitempty"`
+	KeyAlgorithm util.CertKeyAlgorithm `json:"keyAlgorithm,omitempty"`
+	CertType     util.SshCertType      `json:"certType,omitempty"`
 	Principals   []string                `json:"principals"`
 	TTL          string                  `json:"ttl,omitempty"`
 	KeyID        string                  `json:"keyId,omitempty"`
@@ -23,8 +23,8 @@ type SignSshPublicKeyV1Response struct {
 type IssueSshCredsV1Request struct {
 	ProjectID    string                  `json:"projectId"`
 	TemplateName string                  `json:"templateName"`
-	KeyAlgorithm models.CertKeyAlgorithm `json:"keyAlgorithm,omitempty"`
-	CertType     models.SshCertType      `json:"certType,omitempty"`
+	KeyAlgorithm util.CertKeyAlgorithm `json:"keyAlgorithm,omitempty"`
+	CertType     util.SshCertType      `json:"certType,omitempty"`
 	Principals   []string                `json:"principals"`
 	TTL          string                  `json:"ttl,omitempty"`
 	KeyID        string                  `json:"keyId,omitempty"`
@@ -35,5 +35,5 @@ type IssueSshCredsV1Response struct {
 	SignedKey    string                  `json:"signedKey"`
 	PrivateKey   string                  `json:"privateKey"`
 	PublicKey    string                  `json:"publicKey"`
-	KeyAlgorithm models.CertKeyAlgorithm `json:"keyAlgorithm"`
+	KeyAlgorithm util.CertKeyAlgorithm `json:"keyAlgorithm"`
 }
